tests/wpt/tests/credential-management/fedcm-multi-idp/single-get-after-onload.https.html

<!DOCTYPE html>
<title>Federated Credential Management API multi IDP single get after onload test.</title>
<link rel="help" href="https://fedidcg.github.io/FedCM">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<body>

<script type="module">
import {set_fedcm_cookie,
        request_options_with_mediation_required,
        fedcm_get_and_select_first_account} from '../support/fedcm-helper.sub.js';

const window_loaded = new Promise(resolve => {
  window.addEventListener('load', () => {
    resolve();
  });
});

promise_test(async t => {
  await set_fedcm_cookie();
  await window_loaded;
  const cred = await fedcm_get_and_select_first_account(t, request_options_with_mediation_required());
  assert_equals(cred.token, "token");
}, "Single `get` call after onload is allowed.");

</script>
